Reduce json editors to one JsonObjectEditor
