+++ dc-cudami-admin/src/test/java/de/digitalcollections/cudami/admin/controller/MainControllerTest.java
@@ -6,10 +6,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import de.digitalcollections.cudami.admin.business.api.service.security.UserService;
-import de.digitalcollections.cudami.admin.propertyeditor.LocalizedStructuredContentEditor;
-import de.digitalcollections.cudami.admin.propertyeditor.LocalizedTextEditor;
 import de.digitalcollections.cudami.admin.propertyeditor.RoleEditor;
-import de.digitalcollections.cudami.admin.propertyeditor.StructuredContentEditor;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
@@ -20,14 +17,9 @@ import org.springframework.test.web.servlet.MockMvc;
 @WebMvcTest(MainController.class)
 public class MainControllerTest {
 
-  @MockBean private LocalizedStructuredContentEditor localizedStructuredContentEditor;
-  @MockBean private LocalizedTextEditor localizedTextEditor;
   @Autowired private MockMvc mockMvc;
   @MockBean private RoleEditor roleEditor;
-  @MockBean private StructuredContentEditor structuredContentEditor;
-
   @MockBean private UserDetailsService userDetailsService;
-
   @MockBean private UserService userService;
 
   @Test
